Fix test isolation: mock IPatchingService and IGameLaunchingStrategy
- ApplyPatchCommandHandlerTests now mock IPatchingService (test
  delegation only, not PatchingService internals).
- New PatchingServiceTests test PatchingService directly.
- GameLaunchingCommandHandlerTests mock IGameLaunchingStrategy
  instead of constructing real strategies with mocked dependencies.
